repo ready!

+++ voting-service/internal/entity/Log.go
@@ -6,7 +6,7 @@ type Log struct {
 	ID        int64
 	UserID    int64
 	Action    string
-	PollID    int
-	OptionID  int
+	PollID    int64
+	OptionID  int64
 	CreatedAt time.Time
 }
